Login and Register components moved to pages folder
