+++ client/src/components/App.js
@@ -6,7 +6,6 @@ import Navbar from "./Navbar";
 import Footer from "./Footer";
 import Wrapper from "./Wrapper";
 
-
 function App() {
 
   return (
