+++ client/src/components/Wrapper/index.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Route } from 'react-router-dom';
 
 //COMPONENTS
-import Login from '../Login';
-import Register from '../Register';
+import Login from '../../pages/Login';
+import Register from '../../pages/Register';
 
 // STYLESHEETS
 import "./wrapper.css";
